+++ lib/array/nest/method.ts
@@ -1,0 +1,5 @@
+const method = (arr: any[], id: any = null, link: string = "parent_id"): any[] =>
+  arr.filter((item) => item[link] === id)
+    .map((item) => ({ ...item, children: method(arr, item.id) }));
+
+export = method;