+++ lib/array/whereIn/method.ts
@@ -1,3 +1,5 @@
+import { filter } from "../utils";
+
 const method = (arr: any[], field: string | any[], value?: any[]) => {
     if (value === undefined) {
         value = field as any[];
@@ -6,14 +8,7 @@ const method = (arr: any[], field: string | any[], value?: any[]) => {
 
     const iterator = (item: any) => (value as any[]).indexOf(item) !== -1;
 
-    if (field) {
-        const keys = (field as string).split(".");
-        const reducer = (item: any) => keys.reduce((prev, curr) => prev[curr], item);
-
-        return arr.filter((item) => iterator(reducer(item)));
-    }
-
-    return arr.filter(iterator);
+    return filter(arr, field as string, iterator);
 };
 
 export = method;