+++ lib/object/index.ts
@@ -1,6 +1,9 @@
 /** @namespace Object */
 
 import "./camelCaseKeys";
+import "./clone";
+import "./equals";
+import "./get";
 import "./invert";
 import "./isInstance";
 import "./kebabCaseKeys";
@@ -8,5 +11,6 @@ import "./lowerCaseKeys";
 import "./map";
 import "./mapKeys";
 import "./merge";
+import "./omit";
 import "./size";
 import "./snakeCaseKeys";